ai copilot introduced
The interface has been simplified and a scrollable left bar was created to include all the buttons.
new buttons to import images onto the canvas.
new button to import an image as background
button to download the canvas as image
the animate object parameters now are automatically filled when user presses somewhere